+++ python/tests/source_events_test.py
@@ -1,31 +1,15 @@
-from os import path
+from pathlib import Path
 
 from gherkin.stream.source_events import source_event
 
 
 def test_should_resolve_classic_feature():
-    actual = source_event(
-        path.join(
-            path.dirname(__file__),
-            "..",
-            "..",
-            "testdata",
-            "good",
-            "minimal.feature",
-        ),
-    )
+    actual = source_event(Path(__file__).parents[2] / "testdata/good/minimal.feature")
     assert actual["source"]["mediaType"] == "text/x.cucumber.gherkin+plain"
 
 
 def test_should_resolve_markdown_feature():
     actual = source_event(
-        path.join(
-            path.dirname(__file__),
-            "..",
-            "..",
-            "testdata",
-            "good",
-            "minimal.feature.md",
-        ),
+        Path(__file__).parents[2] / "testdata/good/minimal.feature.md",
     )
     assert actual["source"]["mediaType"] == "text/x.cucumber.gherkin+markdown"